// app/dashboard/components/sub-components/event-components/types.ts
export type Event = {
    id: string;
    user_id: string;
    title: string;
    description: string;
    event_date: string;
    start_time: string;
    end_time: string;
    location: string;
    city: string;
    organizer_name: string;
    contact_email: string;
    contact_phone: string;
    website: string;
    created_at: string;
};